Write state file atomically
If you have two different calls of this script on the same file, one of
the runs may be at the point of "truncate" when writing the state
file. If the other run is trying to read, that will raise and the
default is rescue with 0. This means that your check will bounce between
OK and critical/warning depending if these runs happen to stomp one
another.